requirements.txt
# Core libraries
requests
pydantic
python-dotenv

# DataFrame API
pandas

# CLI
fire==0.6.0

# AWS
boto3==1.34.131

# Reddit API
praw==7.7.1
prawcore==2.4.0

# Google BigQuery
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-bigquery==3.24.0
google-cloud-bigquery-storage==2.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1

# dbt core and adapters
dbt-adapters==1.2.1
dbt-common==1.3.0
dbt-core==1.8.2
dbt-duckdb==1.8.1
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.5.1

# DuckDB
duckdb==1.0.0

# Others
zstandard==0.22.0
numpy=1.26.4

# Development
ruff
pytest
pytest-cov